+++ src/main/java/com/devhoss/repository/IClienteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.devhoss.model.Cliente;
 
 @Repository
-public interface IClienteRepository extends JpaRepository<Cliente, Integer>{
+public interface IClienteRepository extends JpaRepository<Cliente, Long>{
 
 }
